flask-server/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
faiss-cpu==1.9.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.3.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
